refactor: move the selection vocabulary under screen/
`SelectionRange`, `SelectionGeometry`, `SelectionKind`, and `CellSide`
move from the crate root (`selection.rs`) to `screen/selection.rs`.
A selection is owned per `Screen`, so its vocabulary belongs next to
`screen/viewport.rs` and `screen/cursor.rs`, which already keep their
public types beside the state that uses them. The move also frees the
`selection` module name for the crate-private state that the selection
capability will add to the same file, instead of two modules sharing
that name at different depths.

No type or behaviour changes; the prelude keeps re-exporting the same
names.
