+++ crates/orzma_vt/src/lib.rs
@@ -21,7 +21,6 @@ mod hyperlink;
 mod interpreter;
 mod placement;
 mod screen;
-mod selection;
 mod vi;
 
 /// The crate's vocabulary, gathered for downstream consumers.
@@ -40,8 +39,10 @@ pub mod prelude {
     pub use crate::screen::grid::coords::{GridColumn, GridLine, GridPoint, ScreenLine};
     pub use crate::screen::grid::row::Row;
     pub use crate::screen::grid::run::{Run, Style};
+    pub use crate::screen::selection::{
+        CellSide, SelectionGeometry, SelectionKind, SelectionRange,
+    };
     pub use crate::screen::viewport::{DisplayOffset, Scroll, ViewportLine};
-    pub use crate::selection::{CellSide, SelectionGeometry, SelectionKind, SelectionRange};
     pub use crate::vi::{ViCursor, ViModeSwitch};
     pub use crate::{InterpretOutput, OrzmaVt, ResizeChanged, Vt, VtSignal};
 }
