+++ crates/orzma_vt/src/screen/selection.rs
